requirements.txt
streamlit==1.52.1
pandas==2.3.3
pillow==12.0.0
